requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.0.0
Django==4.2.11
hyperlink==21.0.0
idna==3.7
imap-tools==1.6.0
IMAPClient==3.0.1
incremental==22.10.0
isort==5.13.2
msgpack==1.0.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dotenv==1.0.1
redis==5.0.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
zope.interface==6.3
